Add method to get reference numbers to `SchoolService`
Unlike the repository method, the service-level method always returns a
result - if the repository returns `null`, a
`SchoolReferenceNumbersServiceModel` containing only the URN is
returned.

Otherwise, the LAESTAB and UKPRN are returned if they are available; the
LAESTAB is considered to be valid if both the LA Code and Establishment
Number are available, otherwise it will be `null`.