Update `OfstedRepository` to return short inspection data
This adds the `GetOfstedShortInspectionAsync` method to the Ofsted
repository, which returns an instance of the new `OfstedShortInspection`
record.

This method avoids a lot of the complexity involved in getting full
inspection data, mainly walking up the chain of previous URNs. If it
turns out that short inspection data seems to be missing when it should
be present, then it's likely that this method will need to be updated to
also do this.

Rename short inspection