Add method to get Ofsted ratings for a single school to `OfstedRepository`
This adds the `GetOfstedInspectionHistorySummaryAsync` method to the
Ofsted repository, which returns an `OfstedInspectionHistorySummary`
record.

The `OfstedInspectionHistorySummary` record is a container for the
current and previous inspections, represented as instances of the
`OfstedFullInspectionSummary` record. This record contains an optional
`InspectionDate` and the `InspectionOutcome` as an `OfstedRatingScore`.

Like `GetOfstedShortInspectionAsync`, this method skips a lot of the
complexity involved with getting full Ofsted data for a trust, so if
information appears to be missing then this method will also need to be
amended to walk up the chain of previous URNs.