Provide data export service for school Ofsted data
This provides the `OfstedSchoolDataExportService`, which will allow
users to download Excel files containing Ofsted report data for an
individual school.

Currently, this just has a skeleton implementation that matches the data
that will be provided on the 'single headline grades' page, but will be
updated to be more complete in future tickets.