Add links to academies in trust pages
This updates the following pages so that academy names are links to the
overview page for the academy:
- /trusts/academies/in-trust/details
- /trusts/academies/in-trust/pupil-numbers
- /trusts/academies/in-trust/free-school-meals
- /trusts/academies/pipeline/pre-advisory-board
- /trusts/academies/pipeline/post-advisory-board
- /trusts/academies/pipeline/free-schools
- /trusts/ofsted/single-headline-grades

It's worth noting that there are no free schools in the pipeline stage
that have URNs, so in practice this page is unchanged. However, if such
a school exists in the future, this page will also provide a link.

Additionally, for the Ofsted pages, only the single headline grades
should have links; the other pages should just have the academy names as
regular text, and have been unchanged.