Improve display of short inspection outcomes
Currently, short inspection outcomes are free text. This results in
multiple instances where a grade is given with some additional
information, such as "School remains Good (improving) - S5 Next".

Users have identified that they don't need this additional information
and would prefer to see "School remains Good (improving)" in this case.
This change introduces handling for these cases to strip this unneeded
extra information.

Additionally, in situations where a short inspection date exists but the
outcome is `null`, the page currently displays nothing. This change
introduces a check for this case so that "Not available" is displayed
instead.

Otherwise, the short inspection grade is displayed the same as it was
previously.