Use new governance calculation
Using the new `TrustGovernanceRepository` as the source for governors
simplifies the calculation, as the new expectation is that only trustees
and the chair of trustees should be included.

Previously, this was difficult, as there was no way to get the chair
specifically, but now that the repository gives a flat list of
`Governor`s simply filtering by the desired roles gives all of the
governors eligible to be part of the turnover rate calculation.