Create data model for Compare census population statistics
This introduces new datatypes to represent census data from Compare
school and college performance in England.

These types are:

`Statistic<T>`: A wrapper for a value that provides multiple reasons for
a value to be missing. These are:
- `Suppressed`: when publishing data for the measure would produce a
  high probability of uniquely identifying students. This is recorded in
  the database with a value of `SUPP`.
- `NotPublished`: Certain kinds of schools such as independent schools
  have exemptions for some statistics. This is recorded in the database
  with a value of `NP`.
- `NotApplicable`: When the data received from Compare does not include
  a given statistic, or the statistic is not applicable to that school.
  This is recorded in the database with a value of `NA`.
- `NotAvailable`: When no data exists in the database for this
  statistic. This occurs when either the column in the database is
  `null` or no row exists for the school for that year.
- `NotYetSubmitted`: Represents a future statistic that does not yet
  exist.

`CensusYear`: a wrapper for `int` that provides utility methods for
getting the current census year from an `IDateTimeProvider` as well as
past/future census years, respecting that a year is not considered
current until the autumn census is published, which must happen by
October 31st.

`AnnualStatistics<T>`: A `Dictionary<CensusYear, T>`. This will be used
by the repository to represent annually published data. It allows
addressing elements using either an `int` or `CurrentYear` as a key.

`SchoolPopulation`: A record representing a single year's worth of data
to be presented on the "Population" page in the "Pupils" area for a
school. This mostly corresponds to the data stored on the
`edperf_mstr.edperf_fiat` table.