Add repository for pupil census data
This introduces a `PupilCensusRepository` which will provide the data
to be displayed in the "Pupils" area for a school.

The `GetSchoolPopulationStatisticsAsync` method doesn't perform any
filtering on the year, even though the page will only display four
years' worth of data. This is because the year is stored in the format
'2025-2026', which would be unnecessarily cumbersome to parse within SQL
(and by extension in Entity Framework). Therefore, the entire result is
loaded into memory for the service layer to filter as needed. Given that
only one row per school per year should exist, this should not become a
problem within my lifetime.