Add `SchoolPupilService`
This provides a `GetSchoolPopulationStatisticsAsync` method which
filters and shapes the output of the repository method to consistently
supply data of the correct shape, even when the database may have
incomplete information. For example, a school that has only opened this
academic year will have only one census, but four years of data may be
requested for it; in this case, missing years have an `Unknown` value.