Fix accessibility issue in the 'About the data we use' page
This page was introduced in #919
(https://github.com/DFE-Digital/find-information-about-schools-and-trusts/pull/919),
but the heading within the banner is not part of any accessibility
landmark, which causes problems for screenreaders.

This banner heading was introduced to the `_ContentLayout` partial in
order to align the page visually with how it was presented in the
prototype. The 'main' landmark is not part of this partial so any text
rendered here will not be visible.

However, rendering the banner within the 'main' landmark causes it to
render with visible margins.

Therefore, the page has had the banner removed and the heading is now
rendered as plain text within the 'main' landmark, which is more
consistent with the other content pages and more suitable for
accessibility tools.