Introduce separate data sources for population and attendance data
The population data from Compare school and college performance in
England is ingested at a different time than the attendance data, as
they are derived from two different censuses. This replaces the
`CompareSchoolCollegePerformanceEngland` data source with one for each
of these two sets of data, using `MetaCensusIngestionDateTime` and
`MetaAbsenceIngestionDateTime` as the last updated date respectively.