Add `TrustPupilService` class
This service provides two methods:
- `GetTotalPupilCountForTrustAsync`
- `GetPupilCountsForSchoolsInTrustAsync`

Both of these methods use the
`GetMostRecentPopulationStatisticsForTrustAsync` method on the
`PupilCensusRepository`.

The `GetPupilCountsForSchoolsInTrustAsync` method performs a
straightforward transformation on the result of the repository call,
using the `PupilsOnRole` property of the `SchoolPopulation` for each
URN. This allows for any schools where this statistic does not have a
value to propagate this information to the caller.

The `GetTotalPupilCountForTrustAsync` method, on the other hand, acts as
an aggregate, and some value should always be returned. Therefore, any
schools where the `PupilsOnRole` statistic does not have a value are
considered to have 0 as their value when performing the sum - meaning
that the number returned by this method is a lower bound.