Fix division by zero error when calculating free school meal percentages
In rare cases, the number of pupils on role in a school can be zero.
Currently, when this is the case, attempting to calculate the percentage
of pupils eligible for free school meals throws a
`DivideByZeroException`, causing a 500 response to be sent to the user.

In this case, returning *something* is better than causing a
user-visible error, so when the number of pupils on role is zero, the
percentage is defined as also being zero.